fix(process): silence 32-bit narrowing warnings in process_control_server
request.size is a uint64_t (the cross-process control protocol is
bitness-agnostic), but std::vector's constructor and transfer_status() both
take size_t, which is 32 bits on the Windows x86 build - MSVC's /W4 /WX
turns the resulting C4244 into a hard error there. Cast explicitly at the
point of use, matching the existing static_cast<size_t> already used a few
lines below in the same file for the same reason.

+++ src/windows-emulator/process_control_server.cpp
@@ -16,12 +16,12 @@ namespace sogen
     {
         void execute_read_memory(windows_emulator& target, const process_control_request& request, process_control_response& response)
         {
-            std::vector<std::byte> data(request.size);
+            std::vector<std::byte> data(static_cast<size_t>(request.size));
             const auto transferred = copy_guest_range_out(target.memory, request.address, data);
             data.resize(transferred);
 
             response.payload = std::move(data);
-            response.status = transfer_status(transferred, request.size);
+            response.status = transfer_status(transferred, static_cast<size_t>(request.size));
         }
 
         void execute_write_memory(windows_emulator& target, const process_control_request& request, process_control_response& response)
